refactor move search entity type
